Enum fmt_macros::Count

pub enum Count<'a> {
    CountIs(usize),
    CountIsName(&'a str),
    CountIsParam(usize),
    CountIsNextParam,
    CountImplied,
}

Unstable (rustc_private)

A count is used for the precision and width parameters of an integer, and can reference either an argument or a literal integer.

Variants
CountIs(usize)
Unstable (rustc_private)

The count is specified explicitly.

CountIsName(&'a str)
Unstable (rustc_private)

The count is specified by the argument with the given name.

CountIsParam(usize)
Unstable (rustc_private)

The count is specified by the argument at the given index.

CountIsNextParam
Unstable (rustc_private)

The count is specified by the next parameter.

CountImplied
Unstable (rustc_private)

The count is implied and cannot be explicitly specified.