Enum fmt_macros::Flag

pub enum Flag {
    FlagSignPlus,
    FlagSignMinus,
    FlagAlternate,
    FlagSignAwareZeroPad,
}

Unstable (rustc_private)

Various flags which can be applied to format strings. The meaning of these flags is defined by the formatters themselves.

Variants

FlagSignPlus
Unstable (rustc_private)

A + will be used to denote positive numbers.

FlagSignMinus
Unstable (rustc_private)

A - will be used to denote negative numbers. This is the default.

FlagAlternate
Unstable (rustc_private)

An alternate form will be used for the value. In the case of numbers, this means that the number will be prefixed with the supplied string.

FlagSignAwareZeroPad
Unstable (rustc_private)

For numbers, this means that the number will be padded with zeroes, and the sign (+ or -) will precede them.