Struct fmt_macros::FormatSpec

pub struct FormatSpec<'a> {
    pub fill: Option<char>,
    pub align: Alignment,
    pub flags: u32,
    pub precision: Count<'a>,
    pub width: Count<'a>,
    pub ty: &'a str,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Specification for the formatting of an argument in the format string.

Fields
fill: Option<char>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Optionally specified character to fill alignment with

align: Alignment

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Optionally specified alignment

flags: u32

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Packed version of various flags provided

precision: Count<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The integer precision to use

width: Count<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The string width requested for the resulting format

ty: &'a str

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The descriptor string representing the name of the format desired for this argument, this can be empty or any number of characters, although it is required to be one word.