Crate fmt_macros

Unstable (rustc_private)

Macro support for format strings

These structures are used when parsing format strings for the compiler. Parsing does not happen at runtime: structures of std::fmt::rt are generated instead.

Reexports

pub use self::Piece::*;

pub use self::Position::*;

pub use self::Alignment::*;

pub use self::Flag::*;

pub use self::Count::*;

Structs

Argument 	 [Unstable]
Representation of an argument specification.

FormatSpec 	 [Unstable]
Specification for the formatting of an argument in the format string.

Parser 	 [Unstable]
The parser structure for interpreting the input format string. This is modeled as an iterator over Piece structures to form a stream of tokens being output.

Enums

Alignment 	 [Unstable]
Enum of alignments which are supported.

Count 	 [Unstable]
A count is used for the precision and width parameters of an integer, and can reference either an argument or a literal integer.

Flag 	 [Unstable]
Various flags which can be applied to format strings. The meaning of these flags is defined by the formatters themselves.

Piece 	 [Unstable]
A piece is a portion of the format string which represents the next part to emit. These are emitted as a stream by the Parser class.

Position 	 [Unstable]
Enum describing where an argument for a format can be located.